Dividing  BluetoothDeviceSettingActivity into two Activities
