+++ Opel-Manager/src/com/opel/cmfw/controller/CommController.java
@@ -26,6 +26,11 @@ import android.net.wifi.p2p.WifiP2pDevice;
 import android.net.wifi.p2p.WifiP2pManager;
 import android.util.Log;
 
+import com.opel.cmfw.controller.bluetooth.BluetoothSocketWrapper;
+import com.opel.cmfw.controller.bluetooth.FallbackBluetoothSocket;
+import com.opel.cmfw.controller.bluetooth.FallbackException;
+import com.opel.cmfw.controller.bluetooth.NativeBluetoothSocket;
+
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.DataInputStream;
@@ -36,7 +41,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.lang.reflect.Method;
 import java.net.InetSocketAddress;
 import java.net.NetworkInterface;
 import java.net.Socket;
@@ -765,114 +769,6 @@ public class CommController {
             len = 0;
             offset = 0;
         }
-    }
-
-    // RedCarrottt: Fix Bluetooth connection failure bug (Issue #103)
-    public static class NativeBluetoothSocket implements
-            BluetoothSocketWrapper {
-        private BluetoothSocket socket;
-
-        public NativeBluetoothSocket(BluetoothSocket tmp) {
-            this.socket = tmp;
-        }
-
-        @Override
-        public InputStream getInputStream() throws IOException {
-            return socket.getInputStream();
-        }
-
-        @Override
-        public OutputStream getOutputStream() throws IOException {
-            return socket.getOutputStream();
-        }
-
-        @Override
-        public String getRemoteDeviceName() {
-            return socket.getRemoteDevice().getName();
-        }
-
-        @Override
-        public void connect() throws IOException {
-            socket.connect();
-        }
-
-        @Override
-        public String getRemoteDeviceAddress() {
-            return socket.getRemoteDevice().getAddress();
-        }
-
-        @Override
-        public void close() throws IOException {
-            socket.close();
-        }
-
-        @Override
-        public BluetoothSocket getUnderlyingSocket() {
-            return socket;
-        }
-    }
-
-    public class FallbackBluetoothSocket extends NativeBluetoothSocket {
-        private BluetoothSocket fallbackSocket;
-
-        public FallbackBluetoothSocket(BluetoothSocket tmp) throws
-                FallbackException {
-            super(tmp);
-            try {
-                Class<?> clazz = tmp.getRemoteDevice().getClass();
-                Class<?>[] paramTypes = new Class<?>[]{Integer.TYPE};
-                Method m = clazz.getMethod("createRfcommSocket", paramTypes);
-                Object[] params = new Object[]{Integer.valueOf(1)};
-                fallbackSocket = (BluetoothSocket) m.invoke(tmp
-                        .getRemoteDevice(), params);
-            } catch (Exception e) {
-                throw new FallbackException(e);
-            }
-        }
-
-        @Override
-        public InputStream getInputStream() throws IOException {
-            return fallbackSocket.getInputStream();
-        }
-
-        @Override
-        public OutputStream getOutputStream() throws IOException {
-            return fallbackSocket.getOutputStream();
-        }
-
-        @Override
-        public void connect() throws IOException {
-            fallbackSocket.connect();
-        }
-
-        @Override
-        public void close() throws IOException {
-            fallbackSocket.close();
-        }
-    }
-
-    public static class FallbackException extends Exception {
-        private static final long serialVersionUID = 1L;
-
-        public FallbackException(Exception e) {
-            super(e);
-        }
-    }
-
-    public static interface BluetoothSocketWrapper {
-        InputStream getInputStream() throws IOException;
-
-        OutputStream getOutputStream() throws IOException;
-
-        String getRemoteDeviceName();
-
-        String getRemoteDeviceAddress();
-
-        BluetoothSocket getUnderlyingSocket();
-
-        void connect() throws IOException;
-
-        void close() throws IOException;
     }
 
     // Port
